Organize project structure, separate main logic
create main-logic.js ,and separate main logic from index.js into its own module"
